Refactor: カスタムフックの責務を減らした(#35)
