+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=31cff345c34d60f477a3",
+    "/js/app.js": "/js/app.js?id=ba00a1757cbf9577aa9a",
     "/css/app.css": "/css/app.css?id=f4d8b4dcea3280933ba3"
 }
