+++ src/resources/ts/containers/organisms/Header.tsx
@@ -1,13 +1,24 @@
-import React, { FC } from 'react';
+import React, { FC, useCallback } from 'react';
+import { useHistory } from 'react-router-dom';
 import { useQueryClient } from 'react-query';
 import Header from '../../components/organisms/Header';
 import { User } from '../../models/User';
-import useAuth from '../../hooks/useAuth';
+import { useLogout } from '../../hooks/auth';
 
 const EnhancedHeader: FC = () => {
   const queryClient = useQueryClient();
   const user = queryClient.getQueryData('user') as User;
-  const { handleLogout } = useAuth();
+
+  const history = useHistory();
+  const { mutate } = useLogout();
+
+  const handleLogout = useCallback(() => {
+    mutate(undefined, {
+      onSuccess: () => {
+        history.push('/login');
+      },
+    });
+  }, [history, mutate]);
 
   return <Header userName={user?.name} handleLogout={handleLogout} />;
 };
