+++ src/resources/ts/containers/pages/Login.tsx
@@ -1,18 +1,52 @@
-import React, { FC } from 'react';
+import React, { FC, useState, useCallback } from 'react';
+import { useHistory, useLocation } from 'react-router-dom';
 import Login from '../../components/pages/Login';
-import useAuth from '../../hooks/useAuth';
+import { useLogin } from '../../hooks/auth';
 
 const EnhancedLogin: FC = () => {
-  const {
-    email,
-    password,
-    handleChangeEmail,
-    handleChangePassword,
-    handleLogin,
-    loginMutation,
-  } = useAuth();
-  const { error, isLoading } = loginMutation;
+  const history = useHistory();
+  const location = useLocation();
+  const { from } = (location.state as { from: string }) || {
+    from: { pathname: '/' },
+  };
+
+  const { error, isLoading, mutate } = useLogin();
   const statusCode = error?.response?.status;
+
+  const [email, setEmail] = useState('');
+  const [password, serPassword] = useState('');
+
+  const handleChangeEmail = useCallback(
+    (ev: React.ChangeEvent<HTMLInputElement>) => {
+      setEmail(ev.target.value);
+    },
+    []
+  );
+
+  const handleChangePassword = useCallback(
+    (ev: React.ChangeEvent<HTMLInputElement>) => {
+      serPassword(ev.target.value);
+    },
+    []
+  );
+
+  const handleLogin = useCallback(
+    (ev: React.FormEvent<HTMLFormElement>) => {
+      ev.preventDefault();
+      if (!email || !password) {
+        return;
+      }
+      mutate(
+        { email, password },
+        {
+          onSuccess: () => {
+            history.replace(from);
+          },
+        }
+      );
+    },
+    [email, password, history, from, mutate]
+  );
 
   return (
     <Login
